pyproject.toml
[project]
name = "stocks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "requests>=2.33.1",
    "twstock>=1.5.0",
    "yfinance>=1.3.0",
]
